Add class for handing qwiic relays
Adds `relay.QwiicRelay` class to handle qwiic relays.
qwiic_relay.py is taken mostly from SparkFun's qwiic_relay
GitHub repo, with minor modifications. The constructor was
modified not to attempt to create an I2C driver if one is not
provided because this repo does not have the required dependencies
to allow for board-agnostic I2C driver connection and the
Automation2040W isn't in SparkFun's database AFAIK.